So many changes...
* ADDED: Jupyter notebook for comparing runtimes
* UPDATED: example program to reflect naming convention changes admidst
growing pains
* ADDED: Blessed_experiments.py a look at making sure I know how to
render to every cell on the terminal with Blessed
* UPDATED: mock_ups.py with continued exploratory code on how I would
want to interact with the external API. See notes in docstrings for
more specifics.
* ADDED: speed_test_log.csv the store of information for graphing
runtimes
* ADDED: speed_tests.py to test the runtime length of various
dictionary to string conversion methods for the rendering engine
* UPDATED:  __init__ with new imports to pull modules up to the top
level. Thank you David Beazley Modules and Packages: live and let die
* ADJUSTED: moved plie.py to a deprecatied version as most of the
functions and classes in there currently, will take new form, so
getting it out of the way.
* ADDED: Skeleton of the Renderer in renderer.py
* ADDED: some intital tests for Renderer (which at this stage don’t
pass, but that’s commiting mid TDD development stage for you)
* ADDED: text.py as a stand-in for the plie.Text view object that’s
coming.