streamlined lie.Text private methods, added more tests, updated todo list
ADJUSTED: plie.text now has less private methods, without losing
functionality. This does mean a complete rebuild of the screenspace
dict occurs every display, which may or may not be a good idea, but at
this stage it is to early to tell. So I am avoiding premature
optimization.
ADDED: test for center justification in plie.Text