View mockups
ADDED: another view mockup in view_mockups.py exploring how View as a
class might work (instead of just a dict)
ADDED: view_as_class_mockup.py to further flesh out that idea. It is
now the start of the View class. A view class is mostly just a
structural object for aiding in formatting and creation of views.
UPDATED: todo list