Work on specifying what counts as a 'valid view dict'
ADDED: a Backus-Naur grammer for the current Renderer and how it
handles View
ADDED: mockups for upcoming changes to how the renderer will parse View
dicts, including the tentative grammar for it. This creates a more
flexible system, that should allow for more functionality.

note: View as a class may be getting entirely removed / turned into a
very light wrapper around UserDict.